pre barrier slow path
