+++ src/plan/generational/copying/mutator.rs
@@ -1,15 +1,16 @@
 pub(super) use super::super::ALLOCATOR_MAPPING;
 use super::GenCopy;
-use crate::plan::barriers::*;
+use crate::plan::generational::barrier::GenObjectBarrier;
 use crate::plan::generational::create_gen_space_mapping;
 use crate::plan::generational::gc_work::GenNurseryProcessEdges;
 use crate::plan::mutator_context::Mutator;
 use crate::plan::mutator_context::MutatorConfig;
 use crate::plan::AllocationSemantics;
+use crate::scheduler::ProcessEdgesWork;
 use crate::util::alloc::allocators::Allocators;
 use crate::util::alloc::BumpAllocator;
 use crate::util::{VMMutatorThread, VMWorkerThread};
-use crate::vm::{ObjectModel, VMBinding};
+use crate::vm::VMBinding;
 use crate::MMTK;
 
 pub fn gencopy_mutator_prepare<VM: VMBinding>(_mutator: &mut Mutator<VM>, _tls: VMWorkerThread) {
@@ -42,9 +43,10 @@ pub fn create_gencopy_mutator<VM: VMBinding>(
 
     Mutator {
         allocators: Allocators::<VM>::new(mutator_tls, &*mmtk.plan, &config.space_mapping),
-        barrier: Box::new(ObjectBarrier::<GenNurseryProcessEdges<VM>>::new(
+        barrier: Box::new(GenObjectBarrier::new(
             mmtk,
-            *VM::VMObjectModel::GLOBAL_LOG_BIT_SPEC,
+            &gencopy.gen,
+            GenNurseryProcessEdges::<VM>::CAPACITY,
         )),
         mutator_tls,
         config,
