+++ src/plan/generational/global.rs
@@ -1,3 +1,5 @@
+use super::gc_work::ProcessModBuf;
+use crate::plan::generational::gc_work::GenNurseryProcessEdges;
 use crate::plan::global::CommonPlan;
 use crate::plan::ObjectQueue;
 use crate::plan::Plan;
@@ -16,7 +18,9 @@ use crate::util::metadata::side_metadata::SideMetadataSpec;
 use crate::util::options::UnsafeOptionsWrapper;
 use crate::util::ObjectReference;
 use crate::util::VMWorkerThread;
+use crate::vm::ObjectModel;
 use crate::vm::VMBinding;
+use crate::MMTK;
 use std::sync::atomic::AtomicBool;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
@@ -69,6 +73,22 @@ impl<VM: VMBinding> Gen<VM> {
             ),
             gc_full_heap: AtomicBool::default(),
             next_gc_full_heap: AtomicBool::new(false),
+        }
+    }
+
+    pub fn add_barrier_modbuf(&self, mmtk: &'static MMTK<VM>, modbuf: Vec<ObjectReference>) {
+        debug_assert!(
+            !mmtk.scheduler.work_buckets[WorkBucketStage::Final].is_activated(),
+            "{:?}",
+            self as *const _
+        );
+        if !modbuf.is_empty() {
+            mmtk.scheduler.work_buckets[WorkBucketStage::Closure].add(ProcessModBuf::<
+                GenNurseryProcessEdges<VM>,
+            >::new(
+                modbuf,
+                *VM::VMObjectModel::GLOBAL_LOG_BIT_SPEC,
+            ));
         }
     }
