+++ src/plan/mod.rs
@@ -14,7 +14,7 @@
 //! For more about implementing a plan, it is recommended to read the [MMTk tutorial](/docs/tutorial/Tutorial.md).
 
 mod barriers;
-pub use barriers::BarrierSelector;
+pub use barriers::{BarrierSelector, GenObjectBarrier};
 
 pub(crate) mod gc_requester;
