+++ src/util/mod.rs
@@ -52,7 +52,7 @@ pub mod metadata;
 /// Forwarding word in object copying.
 pub(crate) mod object_forwarding;
 /// MMTk command line options.
-pub(crate) mod options;
+pub mod options;
 /// Utilities funcitons for Rust
 pub(crate) mod rust_util;
 /// Sanity checker for GC.
